+++ lib/src/runtime/native_bindings.dart
@@ -146,6 +146,9 @@ external ffi.Pointer<ffi.Char> hfRefJson(
 @ffi.Native<ffi.Pointer<ffi.Char> Function()>(symbol: 'dinf_hf_cache_root')
 external ffi.Pointer<ffi.Char> hfCacheRoot();
 
+@ffi.Native<ffi.Pointer<ffi.Char> Function()>(symbol: 'dinf_hf_token')
+external ffi.Pointer<ffi.Char> hfToken();
+
 @ffi.Native<
   ffi.Pointer<ffi.Char> Function(
     ffi.Pointer<ffi.Char>,